backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Telegram Bot
aiogram==3.3.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-dateutil==2.8.2
pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0

# Web scraping (для парсинга ссылок)
beautifulsoup4==4.12.3
lxml==5.1.0

# Background tasks
celery==5.3.6
flower==2.0.1

# Monitoring
sentry-sdk[fastapi]==1.40.0
